mcp-server/pyproject.toml
[project]
name = "dossier-mcp"
version = "0.1.0"
description = "MCP server for Dossier workflows"
authors = [
    {name = "Liberio AI", email = "team@liberio.ai"}
]
requires-python = ">=3.12"
dependencies = [
    "mcp>=1.0.0",
    "ghapi>=1.0.0",
    "python-frontmatter>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[project.scripts]
dossier-mcp = "server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.pytest.ini_options]
asyncio_mode = "auto"
